Moved comment class to new file
